[AN] fix: 토큰 없을 시 회원탈퇴 막기 #281

+++ android/app/src/main/java/com/onair/hearit/presentation/MainViewModel.kt
@@ -61,9 +61,14 @@ class MainViewModel(
     fun withdraw() {
         viewModelScope.launch {
             val token = dataStoreRepository.getAccessToken().getOrNull()
+            if (token == null) {
+                _toastMessage.value = R.string.withdraw_fail
+                _withdrawState.value = false
+                return@launch
+            }
 
             authRepository
-                .withdraw(token?.toBearerToken())
+                .withdraw("Bearer $token")
                 .onSuccess {
                     UserApiClient.instance.unlink { error ->
                         if (error != null) {
